+++ ConsoleExamples/Examples/CSVPredict/marketevaluate.cs
@@ -33,7 +33,7 @@ namespace Encog.Examples.CSVMarketExample
             loader.GetFile(newfileLoad);
 
             var result = new MarketMLDataSet(loader,
-                                             Config.INPUT_WINDOW, Config.PREDICT_WINDOW);
+                                            (ulong) Config.INPUT_WINDOW, (ulong)Config.PREDICT_WINDOW);
           //  var desc = new MarketDataDescription(Config.TICKER,
                                               //   MarketDataType.Close, true, true);
