+++ ConsoleExamples/Examples/Market/MarketEvaluate.cs
@@ -31,7 +31,7 @@ namespace Encog.Examples.Market
         {
             IMarketLoader loader = new YahooFinanceLoader();
             var result = new MarketMLDataSet(loader,
-                                             Config.INPUT_WINDOW, Config.PREDICT_WINDOW);
+                                             (ulong) Config.INPUT_WINDOW,(ulong) Config.PREDICT_WINDOW);
             var desc = new MarketDataDescription(Config.TICKER,
                                                  MarketDataType.AdjustedClose, true, true);
             result.AddDescription(desc);
