+++ ConsoleExamples/Examples/Predict/PredictSunspot.cs
@@ -121,7 +121,7 @@ namespace Encog.Examples.Predict
                 TemporalDataDescription.Type.Raw, true, true);
             result.AddDescription(desc);
 
-            for (int year = TrainStart; year < TrainEnd; year++)
+            for (ulong year = TrainStart; year < TrainEnd; year++)
             {
                 var point = new TemporalPoint(1) {Sequence = year};
                 point.Data[0] = _normalizedSunspots[year];
